requirements.txt
psycopg[binary]==3.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
